fix(tests): failing unit and integration tests (#167)
* CI self-hosted config test #1

* Fix incorrect duplicated job ID

* CI test #2

* CI Test #3

Results of test #2: integration tests work when using GitHub-hosted Actions latest ubuntu

* CI Test #4

* Test unit/integration tests on hosted runner

Test unit/integration tests on the hosted runner to determine if the cause of the failing tests is related to differences between self-hosted and hosted environment.

* Update CI conditional statements of self-hosted runners

Update CI conditional statements that referenced variables only applicable to self-hosted runners. Change to the equivalent for hosted runners.

* Remove very_good_analysis linting dependency

Flutter is not fetching dependencies for sub-packages in the `packages` directory, so the very_good_analysis files are missing until `flutter pub get` is ran for each package

* Add composite actions for asset generation and build validation

* Remove chromedriver from unit tests workflow

* Add missing `shell` property to the `generate-assets` action

Also remove the default value, and unset the environment variable if it does not exist

* Delete the `build/` directory between builds to clear AssetManifest.bin

Rebuilds do not appear to consistently update AssetManifest.bin, even if there are new assets generated between builds (icons, coin configs)

* Add code coverage report as artefact upload

Temporary until link with 3rd-party sources like CodeCov is setup

* Only fail the coverage step if the report does not exist

The flutter test command failing or passing should not affect the test coverage step success/failure

* Fix profit/loss unit tests

Binance closing price introduces some variability in the calculated profit/loss that has to be accounted for. The margin of error can be reduced by improved price and timestamp matching, but this should suffice for now

* Fix code coverage generation action

I forgot the spaces in the bash condition :(. It should be "if [ ... ]; then"

* Revert to using `strings` before `grep` to be platform independent

macOS handles binary files differently to Linux, meaning that `grep -i ...` works with binary files on Linux systems, but not macOS

* Remove trailing spaces in zip generation command

Trailing spaces after the backslash in a multi-line command causes it to fail with "command not found"

* Prefix artefact with runner name & move coverage step to end of ui tests

* Update job names and add descriptions

* Switch build_transformer over to the new sdk repo

* Remove old build_transformer package

* Revert changes to web/src/index.html

* Add index.html to .gitignore

* Move webpack JS files out of web/src

* Revert changes to index.html once more :(

* Untrack index.html & update build transformer package

* Use default GITHUB_TOKEN to make authenticated requests

* Add test case for failing binance requests

* demo data generator: check if coin is supported before fetching

* Add validation step for build_config.json

* Fix silently failing asset manifest validation check

* Add json validity checks for coin configs

* Update commit sha of build transformer package

* Use abstract repository class in generate demo data

* Install Chrome alongside chromedriver on linux for UI tests

* Skip generate demo test data unit tests

* Bump build transformer to dev branch

* Switch UI and unit test workflows back to self-hosted runners

* Restrict coverage report to linux-based runners

* Improve npm build cross-platform compatibility

'&&' is invalid in Windows PowerShell. ';' works on Windows, macOS, and Linux

* Move Flutter and NodeJS installation to composite action

* Add github token to UI and unit test steps

Seeing 403 responses on self-hosted runners

* Fix UI integration tests on self-hosted macos

* Export logs for safaridriver in workflows

Use the --diagnose flag when starting safaridriver

* Fix failing restore wallet integration test

* Add verbose flag to integration test runner

* Fix issues with pumpAndSettle timeout

Add pumpNFrames and use that instead. This is based on the recommendation in the docs: https://api.flutter.dev/flutter/flutter_test/WidgetTester/pumpAndSettle.html

* Fix missing balance overview and back button

* Fix wallet & wallet manager integration tests

* Fix maker & taker integration tests

* Fix misc integration tests

* Fix analyse warnings and remove support page test

* Pin chromedriver to the self-hosted runner version

116.0.5845.96 with path /opt/google/chrome/chrome

* Use mock binance repository for unit testing

Using the live Binance endpoints resulted in some variance of the closing price & final calculation that caused the unit tests to fail on occasion.

* Add note about the possibility of 403 error

---------

Co-authored-by: CharlVS <77973576+CharlVS@users.noreply.github.com>
Co-authored-by: Francois <11577022+takenagain@users.noreply.github.com>