feat(qa): Add device preview
NB: This branch is intended to track `dev`, but will not be merged into `dev`.

Automated CI builds will be investigated in the future. The build associated with this commit (latest `dev` at time of commit) is available on https://komodo-preview.web.app/wallet and will be manually updated regularly.