Add multi-address support to "Faucet" button (#2533)
* Add multi-address support to "Faucet" button

Remove the faucet button from the coin overview and add it as a button icon in the addresses list

Move references to faucet_button.dart from the details page to the coin address list. Note that you might have to change how the page is called in coin_details.dart to pass the selected address of the coin (rather than defaultAddress)

NB: Take a branch from dev  and name it add/multi-address-faucet

* isMobile not displaying Faucet button

The button would disappear when the browser was minimized to mobile size. Also did some UI changes to make the button look more like swap address.

* Update coin_addresses.dart

* Update faucet_button.dart

* Created Bloc for Faucet Button

Faucet Button now uses Bloc.

Problem: showDialog appears for each addressCard because they each have a faucet button.

* Fixed multiple Dialogs

Fixed multiple dialogs appearing for each faucet button by creating an activeAddress

* Updated manner of keeping Active Address

Made changes of tracking active address to not use a static variable but rather handle it in the bloc and not in the button.

Made some changes to the UI to show unactive buttons.

Improved state management in Listenerwhen.

* UI adjustments

Made changes to Faucet Alert Dialog to be more clear.

* Feedback adjustments

Removed the Faucet cubit as it is no longer necessary, updated state name to align with bloc naming conventions, and adding more clarity to faucet_button for the listenWhen.

* Update faucet_button_state.dart

Changed names to align with Bloc Naming conventions

* Refactored all names

* fix: merge errors

- remove receive CoinPageType and handler
- update _Balance to `dev` version
- add PubkeyReceiveDialog

---------

Co-authored-by: WesleyAButt <89137667+WesleyAButt@users.noreply.github.com>
Co-authored-by: CharlVS <77973576+CharlVS@users.noreply.github.com>
Co-authored-by: Francois <takenagain@users.noreply.github.com>