Aligned Column Headers (#2577)
* Aligned Column Headers

Lined up headers with coin details and fixed extra padding with 24H

* feat: Rework unauthenticated assets list

* Aligned Column Headers

Lined up headers with coin details and fixed extra padding with 24H

* fix: revert changed Firebase site ID

* fix(ui): Fix expanded state issues in assets list when searching

* fix(ui): Fix PR QA review issue and Coderabbit feedback

* Aligned coin list headers

Lined up the headers that it fits the list of coins better by changing the ratio.

* Updated header name for chart

It was showing price instead of chart as the header for the asset list.

* fix: address qa review

* fix: formatting

---------

Co-authored-by: CharlVS <77973576+CharlVS@users.noreply.github.com>