feat: ZHTLC support (#3158)
* fix(screenshot-sensitivity): notify listeners in postFrameCallback

fixes "setState() or markNeedsBuild() called when widget tree was locked. This ScreenshotSensitivity widget cannot be marked as needing to build because the framework is locked."  and "DartError: setState() or markNeedsBuild() called during build.
This ScreenshotSensitivity widget cannot be marked as needing to build because the framework is already in the process of building widgets. A widget can be marked as needing to be built during the build phase only if one of its ancestors is currently building. This exception is allowed because the framework builds parent widgets before children, which means a dirty descendant will always be built. Otherwise, the framework might not visit this widget during this build phase."

* chore(sdk): switch to zhtlc branch

* feat(zhtlc): add configuration and param download dialogs

* refactor: generate localisations, and fix theming on popup dialog

* refactor: use repositoryprovider and  adjust status bar width

* perf(coins-manager): add input debouncer and cache heavy operations

- cache heavy operations that aren't expected to change regularly.
- add debounce on search input

* fix(activation): swap page activation, improve list and disposal management

* fix(zhtlc): update activation warning, add padding, and stealth text

* feat(orderbook): migrate to v2 orderbook RPC via SDK

* perf(dex): remove enabled assets rpc call from taker validator

With ZHTLC enabled, the `get_enabled_coins` RPC takes long - not yet sure why - and the taker form calls the validator regularly, resulting in long loading times and input delays

* style(zhtlc): improve dialog layout, add transaction note, fix hidden spinner

* fix(market-metrics): filter out inactive coins before starting calcs

fetching tx history of activating coins wrecks havok on startup, especially for ARRR

* fix(zhtlc): workaround "Task is finished" error with retry on activation

Activation throws a bunch of uncaught errors, takes, long, but with this the coin should not remain in suspended state when it does finally activate. TODO(takenagain): investigate and fix cause in SDK

* feat(zhtlc): add activation status to banner

* fix(withdraw): add recipient address and memo (if not empty)

* fix(zhtlc): filter out active assets before attempting zhtlc activation

This logic was mistakenly not copied over from the normal coin activation function

* fix(wallet): add resilience to delisted coins for wallet coin metadata

* fix(version-info): ensure that apiversion and commit has persist

* perf(market-data): add update backoff strategy and guard against fetching tx history of unactivated assets (#3162)

* fix(wallet): add resilience to delisted coins for wallet coin metadata

* fix(version-info): ensure that apiversion and commit has persist

* perf(market-metrics): add backoff strategy for periodic updates

TODO: ensure that the portfolio growth event is only fired once to avoid restarting the periodic updates on every user input (likely the case currently)

* refactor(market-metrics): extract common coin filtering for test coin removal

* refactor(market-metrics): extract calculation functions into extension

* perf(coins-bloc): add initialisation status tracking for startup

* refactor: add loop exits, update comments, remove unused event member

* perf(dex): reduce overhead in coin selection and filtering

* fix(dex): only update sell amount on order selection if no value present

also cap to the available sell amount

* fix(zhtlc): track ongoing activations to prevent duplicate requests

* fix(taker-validator): revert to getting active assets from API

getWalletAssets was used to improve form validation speed to fix hitching and lengthy delays, but it is not correct and allowed for form submissions with inactive coins, resulting in unexpected error messages

* chore(sdk): switch back to dev branch (sdk branch merged)