fix(text-input-action): use contextual input actions other than none
none is not supported on iOS and throws errors in debug mode

If a developer chooses an inappropriate [TextInputAction] when running in debug mode, an error will be thrown. If the same thing is done in release mode, then instead of sending the inappropriate value, Android will use "unspecified" on the platform side and iOS will use "default" on the platform side.